microservice/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
Deprecated==1.3.1
distro==1.9.0
fastapi==0.136.1
google-auth==2.50.0
google-genai==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
limits==5.8.0
packaging==26.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-multipart==0.0.27
requests==2.33.1
slowapi==0.1.9
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
websockets==16.0
wrapt==2.1.2
